+++ src/opi/input/blocks/block_output.py
@@ -8,5 +8,9 @@ class BlockOutput(Block):
 
     _name: str = "output"
     # > options
-    jsonpropfile: bool | None = None  # If True, generate JSON property file
-    jsongbwfile: bool | None = None  # If True, generate JSON gbw file
+    jsonpropfile: bool | None = None
+    """If True, generate JSON property file"""
+    jsongbwfile: bool | None = None
+    """If True, generate JSON gbw file."""
+    dumpactints: bool | None = None
+    """Dump the active integrals from a CASSCF calculation in an FCIDUMP file."""